Create temporary directory if it doesn't exist
Try to create temporary directory used to store files. This package
can be used in the context of containers created “FROM scratch” that
lack /tmp directory. Rather than put the burden of taking care about
this on the package consumer, create this directory in the context
of the Handler function call.